requirements.txt
google==3.0.0
google-generativeai==0.8.3
google-cloud-aiplatform
streamlit==1.28.0
gunicorn==20.0.4
python-dotenv==1.0.0